Website/requirements.txt
Django==1.11.8
django-rest-framework==0.1.0
djangorestframework==3.8.2
Pillow==5.1.0
PyJWT==1.6.4
pytz==2018.4
